Changed how dummy modules work again.
