+++ builts/_spaces/_hypercube.py
