+++ builts/_spaces/hypercube.py
