--- builts/perambulator.py
@@ -1,3 +1,0 @@
-from everest.builts._perambulator import Perambulator
-CLASS = Perambulator
-build, get = CLASS.build, CLASS.get
